# api-specs/schack-se-api-docs.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "https://member.schack.se/",
      "description": "Default Server URL"
    }
  ],
  "paths": {
...
    "/public/api/v1/tournamentresults/team/table/id/{id}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta turneringstabell (klubblagturnering) för en grupp i en turnering",
        "operationId": "getTeamResultTable",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamTournamentEndResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/tournamentresults/team/roundresults/id/{id}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta rondresultaten (klubblagturnering) för en grupp i en turnering",
        "operationId": "getTeamRoundResults",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentRoundResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/tournamentresults/team/roundresults/id/{id}/memberid/{memberid}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta rondresultaten för en spelare i en turnering",
        "operationId": "getTeamRoundResults_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "memberid",
            "in": "path",
            "description": "id för medlemmen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentRoundResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/tournamentresults/table/memberid/{id}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta individuella turneringsresultat/anmälningar för en medlem",
        "operationId": "getMemberTableResults",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för medlemmen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentEndResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/tournamentresults/table/id/{id}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta turneringstabell (endast individuell än så länge) för en grupp i en turnering",
        "operationId": "getResultTable",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentEndResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/tournamentresults/roundresults/id/{id}": {
      "get": {
        "tags": [
          "tournament-results-public-controller"
        ],
        "summary": "Hämta rondresultaten (endast individuell än så länge) för en grupp i en turnering",
        "operationId": "getRoundResults",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TournamentRoundResultDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/public/api/v1/tournament/tournament/id/{id}": {
      "get": {
        "tags": [
          "tournament-structure-controller"
        ],
        "summary": "Hämta turneringsinfo givet en turnering",
        "operationId": "getTournament",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för turneringen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/api/v1/tournament/group/id/{id}": {
      "get": {
        "tags": [
          "tournament-structure-controller"
        ],
        "summary": "Hämta turneringsinfo givet en grupp i en turnering",
        "operationId": "getTournamentFromGroup",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för gruppen",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/api/v1/tournament/class/id/{id}": {
      "get": {
        "tags": [
          "tournament-structure-controller"
        ],
        "summary": "Hämta turneringsinfo givet en klass/division i en turnering",
        "operationId": "getTournamentFromClass",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för klass/division",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentDto"
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/ratinglist/federation/date/{ratingdate}/ratingtype/{ratingtype}/category/{category}": {
      "get": {
        "tags": [
          "rating-list-public-controller"
        ],
        "summary": "Hämta rankinglista för hela SSF, givet datum, rankingkategori och medlemskategori",
        "operationId": "getFederation",
        "parameters": [
          {
            "name": "ratingdate",
            "in": "path",
            "description": "datum för rankinglistan YYYY-MM-DD",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ratingtype",
            "in": "path",
            "description": "rankingtyp, Långparti=1,Snabb=6,Blixt=7",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "category",
            "in": "path",
            "description": "Medlemskategori: Alla=0, Damer=5, Juniorer=1, Kadetter=2, Miniorer=6, Knattar=7, Veteraner=4",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerInfoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/ratinglist/district/{id}/date/{ratingdate}/ratingtype/{ratingtype}/category/{category}": {
      "get": {
        "tags": [
          "rating-list-public-controller"
        ],
        "summary": "Hämta rankinglista för ett distrikt, givet datum, rankingkategori och medlemskategori",
        "operationId": "getDistrictRatingList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för distriktet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ratingdate",
            "in": "path",
            "description": "datum för rankinglistan YYYY-MM-DD",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ratingtype",
            "in": "path",
            "description": "rankingtyp, Långparti=1,Snabb=6,Blixt=7",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "category",
            "in": "path",
            "description": "Medlemskategori: Alla=0, Damer=5, Juniorer=1, Kadetter=2, Miniorer=6, Knattar=7, Veteraner=4",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerInfoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/ratinglist/club/{id}/date/{ratingdate}/ratingtype/{ratingtype}/category/{category}": {
      "get": {
        "tags": [
          "rating-list-public-controller"
        ],
        "summary": "Hämta rankinglista för en klubb, givet datum, rankingkategori och medlemskategori",
        "operationId": "getClubRatingList",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för klubben",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ratingdate",
            "in": "path",
            "description": "datum för rankinglistan YYYY-MM-DD",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ratingtype",
            "in": "path",
            "description": "rankingtyp, Långparti=1,Snabb=6,Blixt=7",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "category",
            "in": "path",
            "description": "Medlemskategori: Alla=0, Damer=5, Juniorer=1, Kadetter=2, Miniorer=6, Knattar=7, Veteraner=4",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerInfoDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/player/{id}/date/{date}": {
      "get": {
        "tags": [
          "player-public-controller"
        ],
        "summary": "Hämta spelarinfo",
        "operationId": "getPlayer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id för spelare",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "date",
            "in": "path",
            "description": "datum för spelarinfo (klubbtillhörighet, ranking, etc) YYYY-MM-DD",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerInfoDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/api/v1/player/fideid/{id}/date/{date}": {
      "get": {
        "tags": [
          "player-public-controller"
        ],
        "summary": "Hämta spelarinfo givet FIDE-id",
        "operationId": "getPlayerByFIDEId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "FIDE-id för spelare",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "date",
            "in": "path",
            "description": "datum för spelarinfo (klubbtillhörighet, ranking, etc) YYYY-MM-DD",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerInfoDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/public/api/v1/organisation/district/clubs/{districtid}": {
      "get": {
        "tags": [
          "organisation-public-controller"
        ],
        "summary": "Hämta klubbar i distrikt",
        "operationId": "getClubsInDistrict",
        "parameters": [
          {
            "name": "districtid",
            "in": "path",
            "description": "id för distriktet",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClubDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/organisation/club/{clubid}": {
      "get": {
        "tags": [
          "organisation-public-controller"
        ],
        "summary": "Hämta klubb",
        "operationId": "getClub",
        "parameters": [
          {
            "name": "clubid",
            "in": "path",
            "description": "id för klubben",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClubDTO"
                }
              }
            }
          }
        }
      }
    },
    "/public/api/v1/organisation/club/exists/{name}/{id}": {
      "get": {
        "tags": [
          "organisation-public-controller"
        ],
        "summary": "Kontrollera om klubbnamn finns, annat än för klubb-id't som skickas in",
        "operationId": "clubNameExists",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "namn för klubben",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "id för klubben som det frågas för",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "MemberLASKRatingDTO": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PlayerInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthdate": {
            "type": "string",
            "format": "date-time"
          },
          "sex": {
            "type": "integer",
            "format": "int32"
          },
          "fideid": {
            "type": "integer",
            "format": "int32"
          },
          "country": {
            "type": "string"
          },
          "club": {
            "type": "string"
          },
          "clubId": {
            "type": "integer",
            "format": "int32"
          },
          "elo": {
            "$ref": "#/components/schemas/MemberFIDERatingDTO"
          },
          "lask": {
            "$ref": "#/components/schemas/MemberLASKRatingDTO"
          }
        }
      },
...
      "TeamTournamentEndResultDto": {
        "type": "object",
        "properties": {
          "points": {
            "type": "number",
            "format": "float"
          },
          "secPoints": {
            "type": "number",
            "format": "double"
          },
          "place": {
            "type": "integer",
            "format": "int32"
          },
          "contenderId": {
            "type": "integer",
            "format": "int32"
          },
          "teamNumber": {
            "type": "integer",
            "format": "int32"
          },
          "wonGames": {
            "type": "integer",
            "format": "int32"
          },
          "drawGames": {
            "type": "integer",
            "format": "int32"
          },
          "lostGames": {
            "type": "integer",
            "format": "int32"
          },
          "club": {
            "$ref": "#/components/schemas/ClubDTO"
          }
        }
      },
...
      "TournamentRoundResultDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "groupdId": {
            "type": "integer",
            "format": "int32"
          },
          "roundNr": {
            "type": "integer",
            "format": "int32"
          },
          "board": {
            "type": "integer",
            "format": "int32"
          },
          "homeId": {
            "type": "integer",
            "format": "int32"
          },
          "homeTeamNumber": {
            "type": "integer",
            "format": "int32"
          },
          "awayId": {
            "type": "integer",
            "format": "int32"
          },
          "awayTeamNumber": {
            "type": "integer",
            "format": "int32"
          },
          "homeResult": {
            "type": "number",
            "format": "float"
          },
          "awayResult": {
            "type": "number",
            "format": "float"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "finalized": {
            "type": "boolean"
          },
          "publisher": {
            "type": "integer",
            "format": "int32"
          },
          "publishDate": {
            "type": "string",
            "format": "date-time"
          },
          "publishedNote": {
            "type": "string"
          },
          "games": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameDto"
            }
          }
        }
      },
      "TournamentEndResultDto": {
        "type": "object",
        "properties": {
          "points": {
            "type": "number",
            "format": "float"
          },
          "secPoints": {
            "type": "number",
            "format": "double"
          },
          "place": {
            "type": "integer",
            "format": "int32"
          },
          "contenderId": {
            "type": "integer",
            "format": "int32"
          },
          "teamNumber": {
            "type": "integer",
            "format": "int32"
          },
          "wonGames": {
            "type": "integer",
            "format": "int32"
          },
          "drawGames": {
            "type": "integer",
            "format": "int32"
          },
          "lostGames": {
            "type": "integer",
            "format": "int32"
          },
          "groupId": {
            "type": "integer",
            "format": "int32"
          },
          "playerInfo": {
            "$ref": "#/components/schemas/PlayerInfoDto"
          }
        }
      },
...
      "TournamentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "end": {
            "type": "string",
            "format": "date-time"
          },
          "city": {
            "type": "string"
          },
          "arena": {
            "type": "string"
          },
          "type": {
            "type": "integer",
            "format": "int32"
          },
          "ia": {
            "type": "integer",
            "format": "int32"
          },
          "secjudges": {
            "type": "string"
          },
          "thinkingTime": {
            "type": "string"
          },
          "state": {
            "type": "integer",
            "format": "int32"
          },
          "allowForeignPlayers": {
            "type": "integer",
            "format": "int32"
          },
          "teamtournamentPlayerListType": {
            "type": "integer",
            "format": "int32"
          },
          "ageFilter": {
            "type": "integer",
            "format": "int32"
          },
          "nrOfPartLink": {
            "type": "string"
          },
          "orgType": {
            "type": "integer",
            "format": "int32"
          },
          "orgNumber": {
            "type": "integer",
            "format": "int32"
          },
          "ratingRegDate": {
            "type": "string",
            "format": "date-time"
          },
          "ratingRegDate2": {
            "type": "string",
            "format": "date-time"
          },
          "fideregged": {
            "type": "integer",
            "format": "int32"
          },
          "online": {
            "type": "integer",
            "format": "int32"
          },
          "y2cRules": {
            "type": "integer",
            "format": "int32"
          },
          "teamNrOfDaysRegged": {
            "type": "integer",
            "format": "int32"
          },
          "showPublic": {
            "type": "integer",
            "format": "int32"
          },
          "invitationurl": {
            "type": "string"
          },
          "latestUpdated": {
            "type": "string",
            "format": "date-time"
          },
          "secParsedJudges": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "rootClasses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TournamentClassDto"
            }
          }
        }
      },
...
    }
  }
}